Some initial tidying on Boot.Server.ts, though can't make much difference until stateful prerendering is removed
