+++ src/Components/Web.JS/src/BootCommon.ts
@@ -40,7 +40,7 @@ interface BootJsonData {
 
 // Tells you if the script was added without <script src="..." autostart="false"></script>
 export function shouldAutoStart() {
-  return document &&
+  return !!(document &&
     document.currentScript &&
-    document.currentScript.getAttribute('autostart') !== 'false';
+    document.currentScript.getAttribute('autostart') !== 'false');
 }
