+++ src/Components/Web.JS/src/Platform/Circuits/BlazorOptions.ts
@@ -1,0 +1,17 @@
+import { LogLevel } from '../Logging/ILogger';
+
+type SignalRBuilder = (builder: signalR.HubConnectionBuilder) => void;
+
+export interface BlazorOptions {
+  configureSignalR: SignalRBuilder;
+  logLevel: LogLevel;
+}
+
+export function computeCircuitOptions(userOptions?: Partial<BlazorOptions>): BlazorOptions {
+    return { ...defaultOptions, ...userOptions };
+}
+
+const defaultOptions: BlazorOptions = {
+    configureSignalR: (_) => { },
+    logLevel: LogLevel.Warning,
+};
